+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file.
 - Extract a JsonObject and turn it into a JsonModel by with `json.jsonModel("key")`
 - `kotlin-reflect.jar` is now a default dependency. It can be removed if you don't use any of the TableView.`column` functions. Over time, more features will probably require it.
 - Replace View function `UIComponent.replaceWith` now accepts `KClass<View>` and `KClass<Fragment>` as well as UIComponent instances
+- label() and text() builders now accepts an ObservableValue<String> for unidirectional binding
  
 ### Changed